Moved function to its own folder, syled pie chart a bit with hover animation.
