+++ src/components/sets/SetsPage.jsx
@@ -3,9 +3,9 @@ import { Col, Container, Row, Spinner, Form } from "react-bootstrap";
 import { useNavigate } from "react-router-dom";
 import { useEffect, useState } from "react";
 import { useQuery } from "@apollo/client";
-import { GET_SETS, GET_SETS_BY_CODE } from "../../assets/queries.jsx";
-import { ResponsiveLine } from "@nivo/line";
+import { GET_SETS_BY_CODE } from "../../assets/queries.jsx";
 import { ResponsivePie } from "@nivo/pie";
+import { getColorData } from "./SetsFunctions.jsx";
 
 export const SetsPage = ({ setInfo, loadingSets: isLoading, errorSets: loadError }) => {
   const [selectedSet, setSelectedSet] = useState("");
@@ -34,75 +34,16 @@ export const SetsPage = ({ setInfo, loadingSets: isLoading, errorSets: loadError
     }
   };
 
+  //Load data from the API
   const { loading, data } = useQuery(GET_SETS_BY_CODE(selectedSet));
   if (loading) {
     return "Loading...";
   }
-  const { sets } = data || null;
-  let { cards } = {};
-  let latestPrices = {};
-  if (sets && sets.length > 0) {
-    cards = sets[0].cards || {};
 
-    latestPrices = cards.map(function (card) {
-      return card.latestPrice?.price ?? 0;
-    });
-  }
+  //Fetch data from functions
+  const colorData = getColorData(data);
 
-  let colorCount = cards.reduce((colorAccumulator, card) => {
-    if (card.colors.length === 0) {
-      if (colorAccumulator["Colorless"]) {
-        colorAccumulator["Colorless"] += 1;
-      } else {
-        colorAccumulator["Colorless"] = 1;
-      }
-    } else {
-      card.colors.forEach((color) => {
-        if (colorAccumulator[color]) {
-          colorAccumulator[color] += 1;
-        } else {
-          colorAccumulator[color] = 1;
-        }
-      });
-    }
-    return colorAccumulator;
-  }, {});
-
-  console.log(colorCount);
-
-  const colorData = [
-    {
-      id: "Colorless",
-      label: "Colorless",
-      value: colorCount["Colorless"],
-    },
-    {
-      id: "White",
-      label: "White",
-      value: colorCount["W"],
-    },
-    {
-      id: "Blue",
-      label: "Blue",
-      value: colorCount["U"],
-    },
-    {
-      id: "Black",
-      label: "Black",
-      value: colorCount["B"],
-    },
-    {
-      id: "Red",
-      label: "Red",
-      value: colorCount["R"],
-    },
-    {
-      id: "Green",
-      label: "Green",
-      value: colorCount["G"],
-    },
-  ];
-
+  //Page return
   return (
     <>
       <MainNavigation />
@@ -158,6 +99,9 @@ export const SetsPage = ({ setInfo, loadingSets: isLoading, errorSets: loadError
                     data={colorData}
                     margin={{ top: 40, right: 80, bottom: 80, left: 80 }}
                     colors={["#C0C0C0", "#F8F8FF", "#008cff", "#000000", "#ff0000", "#30cc00"]}
+                    cornerRadius={3}
+                    activeOuterRadiusOffset={8}
+                    borderWidth={1}
                   />
                 </div>
               </div>
